add unique constraint on resource, tag
This constraint prevents us from tagging a resource
with the same tag more than once when we're creating
TaggedItems objects via factories.

This is already applied as a soft-constraint at the manager
level when creating via `resource.tags.add()`,
but the factories bypass it. This applies it at the DB level