Merge recent changes from `master` into `main` (#145)
* Feature/readme update (#143)

* docs: update readme

* docs: update based on Linda's (@lpatmo) feedback

* docs: added reference links to spec docs from Slack, updated link to PR127

* docs: backend issue from 129 -> issue 127

Co-authored-by: Linda <lindapnotes@gmail.com>

* Add license (#136)

Can't believe I forgot to do this from the start, heh. We need to make sure we have a license before we apply to https://sentry.io/for/open-source/

* Create FUNDING.yml (#135)

While looking for something unrelated I noticed that it was pretty easy to set up a "sponsor" button for this repo. I kept the original template in there in case we wanted to add things like the github usernames of individual core contributors (up to 4 allowed).

Co-authored-by: Phil <phil@tgrrr.com>